change EditorDraw to float

+++ NewWorld/src/NewWorld/Graphics/EditorDraw.h
@@ -18,50 +18,50 @@ namespace NewWorld::Graphics
 
 		// Actions Local
 	public:
-		static void DrawLine(int x1, int y1, int x2, int y2, const Graphics::Color& color, uint lineWidth = 1);
+		static void DrawLine(float x1, float y1, float x2, float y2, const Graphics::Color& color, uint lineWidth = 1);
 
-		static void DrawRectangle(int x, int y, uint width, uint height, const Graphics::Color& color);
-		static void DrawOutlineRectangle(int x, int y, uint width, uint height, const Graphics::Color& color, uint lineWidth = 1);
+		static void DrawRectangle(float x, float y, float width, float height, const Graphics::Color& color);
+		static void DrawOutlineRectangle(float x, float y, float width, float height, const Graphics::Color& color, uint lineWidth = 1);
 		
-		static void DrawEllipseSlice(int x, int y, uint radiusX, uint radiusY, float angleStart, float angleLength, const Graphics::Color& color, uint verticesCount = 42);
-		static void DrawArc(int x, int y, uint radiusX, uint radiusY, float startAngle, float angleLength, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 42);
+		static void DrawEllipseSlice(float x, float y, float radiusX, float radiusY, float angleStart, float angleLength, const Graphics::Color& color, uint verticesCount = 42);
+		static void DrawArc(float x, float y, float radiusX, float radiusY, float startAngle, float angleLength, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 42);
 
-		static void DrawEllipse(int x, int y, uint radiusX, uint radiusY, const Graphics::Color& color, uint verticesCount = 42);
-		static void DrawOutlineEllipse(int x, int y, uint radiusX, uint radiusY, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 42);
+		static void DrawEllipse(float x, float y, float radiusX, float radiusY, const Graphics::Color& color, uint verticesCount = 42);
+		static void DrawOutlineEllipse(float x, float y, float radiusX, float radiusY, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 42);
 
-		static void DrawOval(int x, int y, uint radius, const Graphics::Color& color, uint verticesCount = 42);
-		static void DrawOutlineOval(int x, int y, uint radius, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 42);
+		static void DrawOval(float x, float y, float radius, const Graphics::Color& color, uint verticesCount = 42);
+		static void DrawOutlineOval(float x, float y, float radius, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 42);
 
-		static void DrawTexture(int x, int y, uint width, uint height, Editor::Assets::Texture& texture);
-		static void DrawTexture(int x, int y, uint width, uint height, Editor::Assets::Texture& texture,
+		static void DrawTexture(float x, float y, float width, float height, Editor::Assets::Texture& texture);
+		static void DrawTexture(float x, float y, float width, float height, Editor::Assets::Texture& texture,
 		uint sampleX, uint sampleY, uint sampleWidth, uint sampleHeight);
 
-		static void DrawTemplateTexture(int x, int y, uint width, uint height, Editor::Assets::Texture& texture,
+		static void DrawTemplateTexture(float x, float y, float width, float height, Editor::Assets::Texture& texture,
 			const Graphics::Color& color);
-		static void DrawTemplateTexture(int x, int y, uint width, uint height, Editor::Assets::Texture& texture,
+		static void DrawTemplateTexture(float x, float y, float width, float height, Editor::Assets::Texture& texture,
 			const Graphics::Color& color, uint sampleX, uint sampleY, uint sampleWidth, uint sampleHeight);
 
-		static void DrawString(int x, int y, String text,
-			const Graphics::Color& color, uint fontSize, uint maxWidth = 0, bool bold = false, bool italic = false);
+		static void DrawString(float x, float y, String text,
+			const Graphics::Color& color, uint fontSize, float maxWidth = 0, bool bold = false, bool italic = false);
 
 		// Actions Global
 	private:
-		static void DrawLine(RawPointer<Editor::EditorWindow> window, int x1, int y1, int x2, int y2, const Graphics::Color& color, uint lineWidth);
+		static void DrawLine(RawPointer<Editor::EditorWindow> window, float x1, float y1, float x2, float y2, const Graphics::Color& color, uint lineWidth);
 
-		static void DrawFillRectangle(RawPointer<Editor::EditorWindow> window, int x, int y, uint width, uint height, const Graphics::Color& color);
-		static void DrawOutlineRectangle(RawPointer<Editor::EditorWindow> window, int x, int y, uint width, uint height, const Graphics::Color& color, uint lineWidth);
+		static void DrawFillRectangle(RawPointer<Editor::EditorWindow> window, float x, float y, float width, float height, const Graphics::Color& color);
+		static void DrawOutlineRectangle(RawPointer<Editor::EditorWindow> window, float x, float y, float width, float height, const Graphics::Color& color, uint lineWidth);
 
-		static void DrawEllipseSlice(RawPointer<Editor::EditorWindow> window, int x, int y, uint radiusX, uint radiusY, float angleStart, float angleLength, const Graphics::Color& color, uint verticesCount = 64);
-		static void DrawArc(RawPointer<Editor::EditorWindow> window, int x, int y, uint radiusX, uint radiusY, float angleStart, float angleLength, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 64);
+		static void DrawEllipseSlice(RawPointer<Editor::EditorWindow> window, float x, float y, float radiusX, float radiusY, float angleStart, float angleLength, const Graphics::Color& color, uint verticesCount = 64);
+		static void DrawArc(RawPointer<Editor::EditorWindow> window, float x, float y, float radiusX, float radiusY, float angleStart, float angleLength, const Graphics::Color& color, uint lineWidth = 1, uint verticesCount = 64);
 		
-		static void DrawTexture(RawPointer<Editor::EditorWindow> window, int x, int y, uint width, uint height,
+		static void DrawTexture(RawPointer<Editor::EditorWindow> window, float x, float y, float width, float height,
 			Editor::Assets::Texture& texture, uint sampleX, uint sampleY, uint sampleWidth, uint sampleHeight);
 		
-		static void DrawTemplateTexture(RawPointer<Editor::EditorWindow> window, int x, int y, uint width, uint height,
+		static void DrawTemplateTexture(RawPointer<Editor::EditorWindow> window, float x, float y, float width, float height,
 			Editor::Assets::Texture& texture, const Graphics::Color& color, uint sampleX, uint sampleY, uint sampleWidth, uint sampleHeight);
 		
-		static void DrawString(RawPointer<Editor::EditorWindow> window, int x, int y, String text,
-			const Graphics::Color& color, uint fontSize, uint maxWidth, bool bold, bool italic);
+		static void DrawString(RawPointer<Editor::EditorWindow> window, float x, float y, String text,
+			const Graphics::Color& color, uint fontSize, float maxWidth, bool bold, bool italic);
 
 		// Utilities
 	private:
